mm/failslab.c: support min-order parameter
It is a common case that allocations fail for high order of contigious
pages such as the allocation failure reported for nfsd_file_cache_init().

For this reason, fail_page_alloc fault injection supports the min-order
parameter.

Add support for the min-order parameter also to failslab fault injection.
With CONFIG_SLAB, setting the min-order value to 5 triggers the error:
 nfsd: unable to allocate nfsd_file_hashtbl

Link: https://lore.kernel.org/linux-nfs/e3cdaeec85a6cfec980e87fc294327c0381c1778.camel@kernel.org/
Signed-off-by: Amir Goldstein <amir73il@gmail.com>