fanotify: preallocate connector and mark outside of group lock
We intend to allow memory shrinkers to evict inodes with fanotify
evictable inode marks.

To avoid getting into direct reclaim with fanotify group lock held,
preallocate the object connector and fanotify mark before adding to
a mark mask and free them if object already has an attached connector
or if another task won the race to create an fanotify mark.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>