fanotify: allow O_PATH flag in event_f_flags
Many of the use cases for fanotify events only use the event->fd
to resolve the path of the event target object.

Using O_PATH for this purpose is more efficient and prevents
exposing a readable fd of the object if that is not required.

To be able to distinguish a user opened O_PATH fd, from fanotify
provided O_PATH event->fd, do not explicitly set FMODE_NONOTIFY
on all O_PATH fds and check explicitly for FMODE_PATH in
fsnotify_file() hook instead to suppress FAN_CLOSE events on close
of O_PATH fds.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>