ovl: use batch iterator for indexdir cleanup
The indexdir cleanup routine is using the helper ovl_dir_read()
to reads the entire content of index directory into a list and then
it verifies the index entries one at a time.

For a large indexdir, this could result in allocating large amounts
of memory for the list at mount time, cache thrashing of filesystem
directory blocks between the first dir iteration that fills the list
and the second list iteration that verifies the directory entries.

The likely case is that the index entries are valid so there is no
modification of the index directory content.

Re-factor the ovl_dir_read() helper into begin/batch/end helpers
and use those helpers in indexdir cleanup routine to limit the size of
the temporary list and the amount of filesystem caches used per batch.

Keep only the index entries that need to be cleaned up in a secondary
list that is iterated at the end to perform the cleanups.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>