ovl: do not scan ovelray watch index on mount
Usually, when ovl has an index dir, it is also used as the "work" dir
where temp files are created, to avoid having to lock three dirs on
copy up to upper dir (work, index and upper dir).

When index dir is used for a read-only watch mount, there is no copy up
to upper dir so locking is less complex and it is ok to use the dedicated
"work" dir for temp files.

By making this change, cleanup of leftover temp files on overlay watch
mount becomes trivial, because the dedicated "work" dir is always removed
and re-created on mount, and there is no need to scan the index dir in
search for leftover temp files, making the mount time O(1) instead of
O(num of index entries).

Signed-off-by: Amir Goldstein <amir73il@gmail.com>