fsnotify: allow the state of attached connector with no marks
There is nothing illegal about the state of attached connector with no
marks, but so far, this state was only temporal before adding the first
mark or removing the last mark.

It could be useful to keep the connector attached after removing the
last mark in the same sense that a memory pool is useful to avoid
frequent allocations in a situation where a single marks is frequently
added and removed to an object.

Adapt fsnotify_destroy_marks() to deal with destroying an object which
has a connector with no marks, by detecting this state and freeing the
connector directly instead of deferring the free to the last
fsnotify_put_mark().

Signed-off-by: Amir Goldstein <amir73il@gmail.com>