fsnotify: support persistent ignore mask in xattr
For groups that opt into persistent ignore masks, check the ignore mask
stored in xattr "security.fsnotify.ignore_mask" before reporting a
permission event on an inode.

To avoid the performance penalty of checking xattr on every permission
event, cache the xattr ignore mask in the inode connector after it was
checked for the first time and keep the connector attached to the inode
even if there are no marks attached until the inode is evicted.

On any permission event, if there is any group subscribed for this event
on the inode, mount or sb, the xattr ignore mask on the inode and the
parent dir will be initialized and consulted before deciding whether or
not to send the event to the group.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>