fanotify: introduce FAN_MARK_SYNC flag
Synchronous add of mark or remove/flush of marks with ignore mask
provides a mothod for safe handover of event handling between two groups:

- First, group A subscribes to some events with FAN_MARK_SYNC
- Then, group B unsubscribes from those events

This method guarantees that any event that both groups subscribed
to will be delivered to either group or to both of them.

Note that FAN_MARK_SYNC provides no synchronization to the object
interest masks, which are checked outside srcu read side.
Therefore, this method does not provide any guarantee regarding
delivery of events which only one of the groups is subscribed to.

For example, if only group B was subscribed to FAN_OPEN_EXEC and only
group A is subscribing only to FAN_OPEN, an execution of a binary file
may not deliever FAN_OPEN_EXEC to group B nor FAN_OPEN to group A.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>