fsnotify: acquire sb write access inside pre modify permission event
For pre modify permission events, acquire sb write access before
leaving SRCU and return >0 to signal that sb write access was acquired.

This can be used to implement safe "handover" of pre modify permission
events between two fanotify groups:

- First, group A subscribes to pre modify events with FAN_MARK_SYNC
- Then, a freeze/thaw cycle is performed on the filesystem
- Finally, group B unsubscribes from those events

This method guarantees that a pre modify event that both groups
subscribed to will be delivered to either group or to both of them.

In case that the pre modify event is delivered only to group B, the
freeze/thaw cycle guarantees that the filesystem modification that
followed that pre modify event was also completed, before the handover
is complete and group B can be closed.

For pre rename permission event, acquire sb write access after the
second of the event pair (i.e. rename to) was authorized.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>