fanotify: allow denying access with errno other than EPERM
There are only two types of response for fanotify permission events
FAN_ALLOW and FAN_DENY.

With FAN_DENY response, user trying to perform the filesystem operation
gets an error with errno set to EPERM.

It would be useful for monitoring process to be able to deny access for
reasons more descriptive than EPERM, for example ETIMEDOUT, when a policy
service is temporary not available.

Permission event listener may set response to FAN_DENY_ERRNO and set
the new field errno in the fanotify_response_error struct to deliver this
errno to the user performing the filesystem operation.

The change in fanotify_response_error struct is backward compatible,
because errno uses high bits in the legacy 32bit response field and old
kernels reject respose value with high bits set.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>