fanotify: introduce class FAN_CLASS_VFS_FILTER
The class FAN_CLASS_VFS_FILTER allows intercepting vfs syscalls with
permission events like the FAN_CLASS_{PRE_,}CONTENT classes, but unlike
those classes it also allows the userspace permission event handler to
make modifications to the filesystem of the object before responding
to the event.

Groups initialized with class FAN_CLASS_VFS_FILTER may only set
"vfs filter" events in their marks masks.

"vfs filter" events are a subset of the permission events which are
delivered to userspace without any vfs locks and restriction, so that
it is safe to make modifications to the filesystem of the object inside
the context of the userspace permission event handler.

The permission events FAN_OPEN*_PERM and FAN_LOOKUP_PERM are classified
as vfs filter events.

FAN_ACCESS_PERM is not classified as vfs filter event, becasue it may be
delivered with sb_start_write() held from write operations with nested
reads such as do_sendfile() and vfs_{copy,clone,dedupe}_file_range().

The async events are not classified as vfs filter events, because they
are not permissions events, but we are going to add pre-vfs filter events
for some of the async events.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>