fanotify: introduce class FAN_CLASS_PRE_VFS
Pre vfs events are a subset of the permission events which are
called without any vfs locks held, so that it is safe to make
modifications to the filesystem of the object inside the context
of permission event handler.

Groups in this class may only set pre vfs events in their mask.

Of the existing four permission events, FAN_ACCESS_PERM may hold
a vfs lock, so it is not allowed in FAN_CLASS_PRE_VFS.

For example, in do_clone_file_range() and in vfs_dedupe_file_range_one(),
remap_verify_area() => security_file_permission(MAY_READ) is called
with sb_start_write() held.

This class has higher priority than FAN_CLASS_PRE_CONTENT.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>