ovl: invalidate dentry with deleted real non-dir
When a non-dir lower or upper is deleted from underneath the overlay
and ovl dentry is still in cache, the real dentry does not become
negative, but unhashed.

Identify this situation and invalidate the ovl dentry in this case.

An unhashed but disconnected real dentry is valid, because it may have
been obtained by opening an nfs file handle.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>