vfs: introduce "start-write-unsafe" file permission hooks
In many of the vfs helpers, security_file_permission() is called before
taking sb_start_write(), but there are exceptions.

The vfs iter_write helpers call permission with sb_start_write() held.

Use the flag MAY_NOT_START_WRITE to annotate the "start-write-unsafe"
permission hooks, so LSM and fsnotify know that calling a nested
sb_start_write() in the permission hook context is not allowed.

This is needed for fanotify "pre content" events.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>