fsnotify: prepare for lookup permission event
We are about to add lookup permission event that would be classified as
a permission event and a read access event.

selinux_path_notify() classifies permission events with an the fsnotify
macro, but open codes classification of read access events.

Define a macro, FSNOTIFY_READ_EVENTS and use it in selinux_path_notify()
to classify the read access events.

The new lookup permission event is not possible on child, so classify
the individual permission events that are possible on child instead of
all permission events as a group.

Use this opportunity to remove the ALL_ prefix from the fsnotify macros
ALL_FSNOTIFY_PERM_EVENTS and ALL_FSNOTIFY_DIRENT_EVENTS - it makes the
macro name longer without adding any information.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>