fanotify: introduce class FAN_CLASS_VFS_FILTER
"vfs filter" events are a subset of the permission events which are
delivered to userspace without any vfs locks, so that it is safe for
usersapce to make modifications to the filesystem of the object inside
the context of permission event handler.

The new class FAN_CLASS_VFS_FILTER is intended to be used by hierarchical
storage management systems (HSM) to intercept "vfs filter" events before
user access to filesystem and populate filesystem objects on-demand.

Groups initialized with class FAN_CLASS_VFS_FILTER may only set
"vfs filter" events in their marks masks.

Currently, there are no events classified as "vfs filter" events.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>