fanotify: introduce FAN_PATH_ACCESS event
The exiting FAN_PRE_ACCESS pre-content event on a directory can be used
to populate a directory before ian application is reading its content.

The new FAN_PATH_ACCESS pre-content event adds the ability to populate
a directory before lookup in the directory index.

The hook is called from lookup_slow() when travesing a directory
during path lookup.  Because it is only relevant for directories,
the FAN_PATH_ACCESS event is generated regardless of FAN_ONDIR flag
in mark mask.

The permission hook cannot be sent to usersapce while lookup is in
RCU walk, so if dcache entries already exist, it is assumed that
FAN_PATH_ACCESS events are not needed, because an event was already
generated when dcache was populated.

To prevent a deadlock, when a directory access monitoring process uses
the event->fd obtained in a FAN_PATH_ACCESS event to lookup a path,
relative to event->fd, this lookup does not generate recursive
FAN_PATH_ACCESS events

Signed-off-by: Amir Goldstein <amir73il@gmail.com>