fanotify: introduce FAN_PATH_ACCESS event
WIP

The exiting FAN_ACCESS_PERM permission event on a directory can be used
to grant permission for reading the directory content.

The new FAN_PATH_ACCESS permission event adds the ability to add
userspace hooks on searching the directory index.

The hook is called from lookup_slow() when travesing a directory
during path lookup.  Because it is only relevant for directories,
the FAN_PATH_ACCESS event is generated regardless of FAN_ONDIR flag
in mark mask.

The permission hook cannot be sent to usersapce while lookup is in
RCU walk, so if dcache entries already exist, it is assumed that
FAN_PATH_ACCESS events are not needed, because an event was already
generated when dcache was populated.

To prevent a deadlock, when a file access monitoring process tries
to read or write the content of a file with the event->fd obtained in
a FAN_ACCESS_PERM permission event, those read/write access do not
generate nested permission events.

Similarly, when the lookup monitoring process uses the event->fd in
a FAN_PATH_ACCESS permission event as the dirfd argument to syscalls,
the lookup of paths relative to dirfd and an openat() using this dfd
do not generate nested permission event.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>