fanotify: introduce FAN_LOOKUP_PERM event
The exiting FAN_ACCESS_PERM permission event on a directory can be used
to grant permission for reading the directory content.

The new FAN_LOOKUP_PERM permission event adds the ability to add
userspace hooks on searching the directory index.

The hook is called from may_lookup() when travesing a directory
during path lookup.  Because it is only relevant for directories,
the FAN_LOOKUP_PERM event is generated regardless of FAN_ONDIR flag
in mark mask.

The permission hook cannot be sent to usersapce while lookup is in
RCU walk, so if any group is handling FAN_LOOKUP_PERM events,
RCU walk will be aborted to re-call the hook in reflink walk.

To prevent a deadlock, when a file access monitoring process tries
to read or write the content of a file with the event->fd obtained in
a FAN_ACCESS_PERM permission event, those read/write access do not
generate nested permission events.

Similarly, when the lookup monitoring process uses the event->fd in
a FAN_LOOUP_PERM permission event as the dirfd argument to syscalls,
the lookup of paths relative to dirfd and an openat() using this dfd
do not generate nested permission event.

Following a regression report from kernel test robot on earlier version
of this patch, we now use an optimized open coded version of the fsnotify
mask checks in lookup_permission(), which checks mnt/sb fsnotify masks
only once after each jump to avoid regressing lookup performance.

Tested-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>