fanotify: introduce FAN_MARK_SYNC flag
Synchronous add/remove of mark mask provides a method for safe
handover of event handling on an object between two groups:

- First, group A subscribes to some events with FAN_MARK_SYNC
- Then, group B unsubscribes from those events and processes
  all the delivered events before shutting down

This method guarantees that any event that both groups subscribed
to on an object will be delivered to either group or to both of them.

In the case of modify events, the "event handling" period includes the
pre-modify event, the filesystem modification that follows it, and the
post-modify async event, so it is also guaranteed that if group A or B
receieve a pre-modify permission event on the marked object, they will
also receive the respective post-modify async event.

FAN_MARK_SYNC must be provided along with either FAN_MARK_ADD or
FAN_MARK_REMOVE, but it may be provided with an empty mask in order to
synchronize a batch of mark mask updates that are all on objects on the
same filesystem.

Note that FAN_MARK_SYNC provides no synchronization to the object's
interest mask, which are checked outside fsnotify_mark_srcu read side.
Therefore, this method does not provide any guarantee regarding the
delivery of access events which only one of the groups is subscribed to.

For example, if only group B was subscribed to FAN_OPEN_EXEC on the mount
and group A is subscribing only to FAN_OPEN on the mount, an execution of
a binary file may not deliever FAN_OPEN_EXEC to group B nor FAN_OPEN to
group A.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>