fsnotify: activate sb write barriers for permission event watchers
Commit e43de7f0862b ("fsnotify: optimize the case of no marks of any type")
optimized the case where there are no fsnotify watchers on any of the
filesystem's objects.

It is quite common for a system to have a single local filesystem and
it is quite common for the system to have some inotify watches on some
config files or directories, so the optimization of no marks at all is
often not in effect.

fanotify permission watchers are far less common, so optimizing the case
of no marks with permission events could improve performance for more
systems, especially after introducing sb write barriers.

Maintain a per-sb counter of objects that have marks with permission
events in their mask and use that counter to optimize out fsnotify
permission hooks.

Pre-modify permission hooks are going to require sb write barrier
support, so activate sb write barriers on sb if there are any permission
event watchers.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>