ovl: add support for "redirect_dir=lazyfollow" on lower only overlay
With mount option "redirect_dir=lazyfollow", if lower data inode
is a metacopy redirect with absolute path, defer the lookup of lower
data inode to the time when the access to data is needed.

"redirect_dir=lazyfollow" is currently only supported on non-upper
mount, because ovl_inode->redirect field is overloaded to store
either upperredirect or the lazyfollow lookup path.

This commit only defers the lookup for later.
Following change will perform the lazy lookup.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>