ovl: deduplicate lowerpath and lowerstack[0]
For the common case of single lower layer, embed the single lower path
in ovl_inode and no external stack allocation is needed.

For directory with more than single lower layer and for regular file
with lowerdata, the lower stack is stored in an external allocation.

Use accessor ovl_lowerstack() to get the embedded or external stack.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>