ovl: replace lowerdata inode reference with lowerdata redirect
Now that we have the entire lower stack in ovl_inode, we do not
need to hold another reference to the lowerdata inode.

Instead, use the vacant ovl_inode space as a place holder for lowerdata
redirect path from the metacopy to lowerdata, which is going to be used
later on for lazy lowerdata lookup.

Use accessors to get the lowerdata path and dentry.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>