fanotify: prepare for reporting fanotify encoded fsid
Adding a mark to a group initialized with FAN_REPORT_FID requires that
the filesystem containing the marked object advertise a non-zero fsid.
This limits the use of FAN_REPORT_FID to a subset of the filesystems.

To make FAN_REPORT_FID avilable on more filesystem, support reporting
"fanotify encoded fsid" -
Instead of the filesystem advertised fsid, when the fsid is zero,
fallback to report the backing device number as fsid.

This change of behavior requires a new init flag FAN_REPORT_ANY_FID,
which is not enabled yet.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>