fanotify: add support for FAN_UNMOUNT event
inotify generates unsolicited IN_UNMOUNT events for every inode
mark before the filesystem containing the inode is shutdown.

Unlike IN_UNMOUNT, FAN_UNMOUNT is an opt-in event that can only be
set on a mount mark and is generated when the mount is unmounted.

FAN_UNMOUNT requires FAN_REPORT_FID and reports an fid info record
with fsid of the filesystem and an empty file handle.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>