ovl: use persistent s_uuid with index=on
With index=on, overlayfs instances are non-migratable, meaning that
the layers cannot be copied without breaking the index.

So when indexdir exists, store a persistent uuid in xattr on the
indexdir to give the overlayfs instance a persistent identifier.

This also makes f_fsid persistent and more reliable for reporting
fid info in fanotify events.

With mount option uuid=nogen, a persistent uuid is not be initialized
on indexdir, but if a persistent uuid already exists, it will be used.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>